Fluent API Adjustments

+++ src/BarberShopAPI/Core/Entities/Catalog.cs
@@ -7,7 +7,7 @@ namespace BarberShopAPI.Core.Entities
         public Clasification Clasification { get; set; }
         public int ClasificationID { get; set; }
         public Employee Employee { get; set; }
-        public int EmployeeID { get; set; }
+        public long EmployeeID { get; set; }
         public bool isEnabled { get; set; }
     }
 }
