+++ src/BarberShopAPI/Core/Entities/Customer.cs
@@ -8,7 +8,7 @@ namespace BarberShopAPI.Core.Entities
         public int customerID { get; set; }
         public DateTime? LastVisit { get; set; }
         public bool IsVIP { get; set; }
-        public int PersonID { get; set; }
+        public long PersonID { get; set; }
         public Person Person { get; set; }
     }
 }
