+++ src/BarberShopAPI/Core/Entities/EmployeeRole.cs
@@ -2,7 +2,7 @@ namespace BarberShopAPI.Core.Entities
 {
     public class EmployeeRole
     {
-        public int EmployeeID { get; set; }
+        public long EmployeeID { get; set; }
         public Employee Employee { get; set; }
         public int RoleID { get; set; }
         public Role Role { get; set; }
