+++ src/BarberShopAPI/Core/Entities/EmployeeSubordinate.cs
@@ -2,9 +2,9 @@ namespace BarberShopAPI.Core.Entities
 {
     public class EmployeeSubordinate
     {
-        public int EmployeeID { get; set; }
+        public long EmployeeID { get; set; }
         public Employee Employee { get; set; }
-        public int subordinateID { get; set; }
+        public long subordinateID { get; set; }
         public Employee Subordinate { get; set; }
     }
 }
